+++ DZ19/src/index.ts
@@ -1,0 +1,4 @@
+export * from './pages/home-page';
+export * from './elements/header';
+export * from './elements/footer';
+export * from './elements/general';
