--- DZ19/src/pages/index.ts
@@ -1,1 +1,0 @@
-export * from './home-page';
